requirements.txt
langchain-openai>=0.1.0
langchain-core>=0.1.0
langgraph>=0.0.20
langchain-community>=0.0.20
faiss-cpu>=1.7.4
streamlit>=1.28.0
pydantic>=2.0.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pillow>=10.0.0
streamlit-javascript>=0.1.5
requests>=2.31.0
streamlit-folium>=0.15.0
folium>=0.14.0
geopy>=2.4.0
shapely>=2.0.0
